requirements.txt
streamlit==1.38.0
torch==2.4.1
torchvision==0.19.1
transformers==4.45.2
Pillow==10.4.0
numpy==1.26.4
h5py==3.11.0